regex : cannot retrieve the value from the 'entities' dictionnary
