--- data/my_data_annotated
@@ -1,1 +1,0 @@
-[('Quel beau cheval!', {'entities': [(10, 15, 'animal')]}), ('Le cheval.', {'entities': [(3, 8, 'animal')]}), ('Cheval blanc, cheval noir.', {'entities': [(0, 5, 'animal')]}), ('LE CHEVAL!', {'entities': [(3, 8, 'animal')]}), ('Monte sur ton cheval le plus noble.', {'entities': [(14, 19, 'animal')]}), ("Le cheval ne vaut pas l'avoine.", {'entities': [(3, 8, 'animal')]}), ('Bride de cheval ne va pas à un âne.', {'entities': [(9, 14, 'animal')]}), ("On connaît le cheval en chemin, et le cavalier à l'auberge.", {'entities': [(14, 19, 'animal')]}), ('A jeune cheval, vieux cavalier.', {'entities': [(8, 13, 'animal')]}), ('A cheval neuf vieux cavalier.', {'entities': [(2, 7, 'animal')]}), ('La mort me tente toujours comme un galop de liberté sur un cheval fou.', {'entities': [(59, 64, 'animal')]}), ("Ce n'est pas toujours le cheval qui gagne l'avoine qui la mange.", {'entities': [(25, 30, 'animal')]}), ("Le meilleur fouet pour faire tirer le cheval, c'est l'avoine.", {'entities': [(38, 43, 'animal')]}), ("J'aime mieux courir après mon cheval que de le relever.", {'entities': [(30, 35, 'animal')]}), ('Un cheval ! un cheval ! Mon royaume pour un cheval !', {'entities': [(3, 8, 'animal')]}), ('Avec un bon cheval, le fouet est inutile.', {'entities': [(12, 17, 'animal')]})]
