+++ data/test
@@ -1,0 +1,2 @@
+Quel beau cheval
+Le cheval.
