+++ data/test_annotated
@@ -1,0 +1,1 @@
+[('Quel beau cheval', {'entities': [(10, 15, 'animal')]}), ('Le cheval.', {'entities': [(3, 8, 'animal')]})]
