+++ python/train_spacy.py
@@ -4,11 +4,4 @@ import random
 from pathlib import Path
 import spacy
 from past.builtins import execfile
-execfile('data_processing.py')
-
-
-filename = "../data/my_data"
-target = ["cheval"]
-label = ["animal"]
-
-#data = create_annoted_data(filename, target, label)
+execfile('data_process.py')
